mantle/platform/aws: expand API to support windows li images
Expand the API to support building Windows License Included images.
The added methods include creating, attaching, detaching, and deleting
volumes, stopping instances, and creating images from instances.

Update RemoveImage to only delete the s3 object if the bucket and path
are specified to allow the method to be used to delete the WinLI image,
does not have an uploaded s3 object associated with it.

A ReplaceRootVolume method is also added to swap the root volume of an
ec2 instance with a target volume.

+++ mantle/platform/api/aws/images.go
@@ -408,22 +408,25 @@ func (a *API) deregisterImageIfExists(name string) error {
 
 // Remove all uploaded data associated with an AMI.
 func (a *API) RemoveImage(name, s3BucketName, s3ObjectPath string) error {
-	err := a.DeleteObject(s3BucketName, s3ObjectPath)
-	if err != nil {
-		if awsErr, ok := err.(awserr.Error); ok {
-			if awsErr.Code() != "NoSuchKey" {
+	// Delete S3 object if bucket and path are specified.
+	if s3BucketName != "" && s3ObjectPath != "" {
+		err := a.DeleteObject(s3BucketName, s3ObjectPath)
+		if err != nil {
+			if awsErr, ok := err.(awserr.Error); ok {
+				if awsErr.Code() != "NoSuchKey" {
+					return err
+				}
+			} else {
 				return err
 			}
 		} else {
-			return err
+			plog.Infof("Deleted existing S3 object bucket:%s path:%s", s3BucketName, s3ObjectPath)
 		}
-	} else {
-		plog.Infof("Deleted existing S3 object bucket:%s path:%s", s3BucketName, s3ObjectPath)
 	}
 
 	// compatibility with old versions of this code, which created HVM
 	// AMIs with an "-hvm" suffix
-	err = a.deregisterImageIfExists(name + "-hvm")
+	err := a.deregisterImageIfExists(name + "-hvm")
 	if err != nil {
 		return err
 	}
@@ -460,17 +463,21 @@ func (a *API) createImage(params *ec2.RegisterImageInput) (string, error) {
 	} else if awserr, ok := err.(awserr.Error); ok && awserr.Code() == "InvalidAMIName.Duplicate" {
 		// The AMI already exists. Get its ID. Due to races, this
 		// may take several attempts.
-		for {
+		timeout := 5 * time.Minute
+		delay := 10 * time.Second
+		err = util.RetryUntilTimeout(timeout, delay, func() error {
 			imageID, err = a.FindImage(*params.Name)
 			if err != nil {
-				return "", err
+				return err
 			}
 			if imageID != "" {
 				plog.Infof("found existing image %v, reusing", imageID)
-				break
+				return nil
 			}
-			plog.Debugf("failed to locate image %q, retrying...", *params.Name)
-			time.Sleep(10 * time.Second)
+			return fmt.Errorf("failed to locate image %q", *params.Name)
+		})
+		if err != nil {
+			return "", fmt.Errorf("error finding duplicate image id: %v", err)
 		}
 	} else {
 		return "", fmt.Errorf("error creating AMI: %v", err)
@@ -495,6 +502,89 @@ func (a *API) createImage(params *ec2.RegisterImageInput) (string, error) {
 	}
 
 	return imageID, nil
+}
+
+// CreateImageFromInstance will return the ImageID of the image created from an ec2 instance
+// A new snapshot is created in the process, so the SnapshotID is also returned
+func (a *API) CreateImageFromInstance(params *ec2.CreateImageInput) (string, string, error) {
+	var imageID string
+	timeout := 5 * time.Minute
+	delay := 10 * time.Second
+
+	res, err := a.ec2.CreateImage(params)
+	if err == nil {
+		imageID = *res.ImageId
+		plog.Infof("created image %v", imageID)
+	} else if awserr, ok := err.(awserr.Error); ok && awserr.Code() == "InvalidAMIName.Duplicate" {
+		// The AMI already exists. Get its ID. Due to races, this
+		// may take several attempts.
+		err = util.RetryUntilTimeout(timeout, delay, func() error {
+			imageID, err = a.FindImage(*params.Name)
+			if err != nil {
+				return err
+			}
+			if imageID != "" {
+				plog.Infof("found existing image %v, reusing", imageID)
+				return nil
+			}
+			return fmt.Errorf("failed to locate image %q", *params.Name)
+		})
+		if err != nil {
+			return "", "", fmt.Errorf("error finding duplicate image id: %v", err)
+		}
+	} else {
+		return "", "", fmt.Errorf("error creating AMI: %v", err)
+	}
+
+	// wait for the AMI to be in the "available" state
+	timeout = 10 * time.Minute
+	err = util.WaitUntilReady(timeout, delay, func() (bool, error) {
+		image, err := a.describeImage(imageID)
+		if err != nil {
+			return false, err
+		}
+
+		if *image.State != ec2.ImageStateAvailable {
+			return false, nil
+		}
+		return true, nil
+	})
+
+	if err != nil {
+		return "", "", fmt.Errorf("waiting for image to be available: %v", err)
+	}
+
+	image, err := a.describeImage(imageID)
+	if err != nil {
+		return "", "", err
+	}
+
+	snapshotID, err := getImageSnapshotID(image)
+	if err != nil {
+		return "", "", err
+	}
+
+	// Attempt to tag inside of a retry loop; AWS eventual consistency means that just because
+	// the FindImage call found the AMI it might not be found by the CreateTags call
+	err = util.RetryConditional(6, 5*time.Second, func(err error) bool {
+		if awserr, ok := err.(awserr.Error); ok && awserr.Code() == "InvalidAMIID.NotFound" {
+			return true
+		}
+		return false
+	}, func() error {
+		// tag the new image and the new snapshot
+		// We do this even in the already-exists path in case the previous
+		// run was interrupted.
+		return a.CreateTags([]string{imageID, snapshotID}, map[string]string{
+			"Name": *params.Name,
+		})
+	})
+	if err != nil {
+		return "", "", fmt.Errorf("couldn't tag image name: %v", err)
+	}
+
+	return imageID, snapshotID, nil
+
 }
 
 // GrantVolumePermission grants permission to access an EC2 snapshot volume (referenced by its snapshot ID)
